Major changes to sentence decomposition and runtime
The algorithms now decompose the requirements into sentences and pieces them back together for the index highlighting to work properly. This means that the regexes now run much quicker, and the algorithms don't hang (freeze) on long or complicated requirements.

Additionally, changes to how the algorithms are called based on the config have also been changed. WIth the new code, there is no need to check the config every sentence (which could involve thousands of checks for large documents). Instead, the config alters the algorithms to be run, and then those algorithms are run over each sentence without the need to check.